requirements.txt
beautifulsoup4
lxml
vaderSentiment
hnswlib==0.8.0
whoosh
fastapi
uvicorn
streamlit
ollama
numpy
python-dateutil
pydantic
PyYAML
langchain-ollama
openai
langchain-core
python-multipart
pytest
pytest-mock
requests
python-dotenv